Add files via upload
The bash script should be downloaded to your CentOS7 which already finished installation and network configuration, your CentOS7 must be able to access to the internet. This script will configure the firewall only open on httpd, if you want to open other ports on firewall you should change the script!